requirements.txt
aiohttp
pillow
piexif
tqdm

